feat(theme): add dark theme
The Dark Theme is now available.
The AxisConfig has now a cleaner styling, splitted by axis or tick styles.
Changing a theme is now easier and it's not necessary tied to a darkmode flag. New
theme can be added and used on the charts.

fix #35

BREAKING CHANGE: The `Theme.AxisConfig` type has a different signature.
It now contains `axisTitleStyle`, `axisLineStyle`, `tickLabelStyle` and
`tickLineStyle` defined as `TextStyle` or `StrokeStyle` elements.
The `Theme` interface is changed in a more flat structure.
`darkMode` prop from `Setting` is removed.
`theme` prop in `Setting` is now a `Theme` type object, not a `PartialTheme`.
You can use `mergeWithDefaultTheme` function to merge an existing theme
with a partial one.