fix(axis labels): offset previous space correctly
This fixes the overlapping labels issue that was surfacing in time series labels where the first and
second labels would overlap even when showOverlappingTicks/Labels was false.  The issue is that the
previousOccupiedSpace variable which was being used to determine if a tick/label should be included
in the set of visible ticks/labels was not being updated correctly for the first tick, causing the
next tick to render (even when it should have been hidden).  This commit fixes that value update as
well as simplifies the initialization of the variable to make the logic flow a bit easier to follow.

fixes #42