feat(interactions): crosshair
Removed any listeners from the canvas elements. Using a map for each geometry on the chart: mappings x value and any relative Y values associated. We invert the mouse coordinates using the xScale to get the right X key to use on that map. This increase the performance on highly/medium dense charts. Two new props are added to the `Settings` specs: `tooltipType` to specify the tooltip type (vertical cursor, crosshair, follow, none) and the `tooltipSnap` to snap the cursor to the grid for linear charts.

close #80, close #58, close #88