ci: add jjbb jobs and utility script
Add the definition for the pull request job so that it can be managed
from this repo. The values in defaults.yml will be merged with the pull
request definition. `check_paths_for_matches` is a utility script used
to prevent unnecessary builds from happening on every push to master due
to problems with the Jenkins `git` plugin.