fix(last_value): compute last value for non stacked series
The percentage stacked mode implies to use the formatted value to be shown in the tooltips and on
the legend. I moved the computation from the splittes series to it's own function after the series
set is formatted. The commit fix the current status where the last value was computed only for
stacked series.