fix: handle splitted series with group value to 0
The current implementation wrongly interpret a zero value in the colorValues list of a legend item
as a single series. This cause that series to be named with the spec name or the spec id. Instead a
splitted series with a group value of zero, should be threated in the same way as anyother splitted
series, using the group value as the series name

fix #288