fix(histogram): fix overflowing annotation with single value
This commit fix an issue where a rect annotation with x0 defined in a single value histogram, is
drawn outisde of the chart.

BREAKING CHANGE: The current coordinate configuration of a rect annotation were inverted. This
commit now reverse them: a rect coordinate with only the x0 value will cover from the x0 value to
the end of the domain, a rect coordinate with only the x1 value will cover the interval from the
beginning of the domain till the x1 value.