fix(rendering): out-of-domain rendering of points/bars/lines/areas (#395)
On a previous PR: https://github.com/elastic/elastic-charts/pull/355 we fixed the clipping of
elements on the chart edges. This introduced an unwanted bug where we where removing from the chart also points that are within the chart range itself. Instead of checking if a geometry position is within the range of the scale, this PR checks for the value to be within the domain of the scale.

fix #386