fix: render stacked bar with stringified values
The stacked bars are computed adding up the previous value with the current one. This works fine if
the passed values are numbers. If the number is codified as a string, the resulting stacked value is
a wrongly concatenated string of values. This commit cast every y value to a number, if NaN or null
it will use null.

fix #487