fix(external pointer): avoid recursive-loops on pointer events
If the consumer code doesn't take care of correctly distribute the pointer events avoiding feeding a
chart with its own cursor position, it will end up with a recursive loop that ultimately throw a max
stack trace. This provide the fix for such case

BREAKING CHANGE: The `onCursorUpdate` Settings property is changed to a more generic
`onPointerUpdate`. The same apply for the event type `CursorEvent` that is now `PointerEvent` and
can assume a `PointerOverEvent` or `PointOutEvent` shape (see TS types)