feat(brush): add multi axis brushing
This commit allow the consumer to configure the direction used for the brush tool. The direction is
by default along the X axis, but can be changed to be along the Y axis or have a rectangular
selection along both axis. For the Y axis values, we are returning an array of values, one for each
Y axis defined (this is determined by the groupId associated with different axis/series groups)

BREAKING CHANGE: The type used by the `BrushEndListener` is now changed to reflect this change. The
new type is on form of `{ x?: [number, number]; y?: Array<{ groupId: GroupId; values: [number,
number]; }> }` where x contains an array of [min,max] values, and the y property is an optional
array of objects, containing the GroupId and the values of the brush for that specific axis.

fix #587