fix(annotations): fix alignment at the edges
This commit fixed the alignment of rect and line annotations near the edge of a data domain. It
aligns the x/y position specified by the dataValues of the annotation to the domain of the chart.

BREAKING CHANGE: In the rectangular annotation, the `y0` parameter of the coordinates now refers to
the minimum value and the `y1` value refers to the maximum value of the y domain. Previously, this
was the opposite.

fix #586