build: remove testing hooks and add random seed  (#684)
Removed the testing from the pre-commit hooks: we are already running the same set of tests on CI and we don't need to run them locally on every commit. Everyone has the ability to run tests independently from the commit.
Run `eslint` using the cache to check only changed files (we should carefully check how this behaves when working on different branches)
Changed also the order of the tasks on the commit hook, now it executes first the `commitlint` to check the commit conventions and then runs the linters. Unfortunately, there is no other way to execute the `commit-msg` hook before the `pre-commit` hook in Husky (typicode/husky#441)
Added the RNG_SEED env variable when running storybook, so we can run VRTs seamlessly on CI and on the local instance (both local versions: `test:integration` and `test:integration:local`).
Removed the yarn.lock file from the shipped release, reducing the npm pack size by half.