fix: missing dash style in line annotation (#692)
This commit re-adds the dash style for line annotations. This probably was an existing feature before we removed Konva library, the story annotations/lines/5_styling.tsx has references to dashWidth and dashGapWidth used on the line style.
The VRT was manually updated as with the current change threshold (0,005% of the whole image) the change is not surfaced. I will open a subsequent PR to check if we can safely remove this threshold

fix #687