fix(partition): linked label on a larger than 180 degree slice (#726)
This commit fix an issue arose when a single large (>180 degrees) slice is forced to have a linked label, eg. via a large or infinite `maximumSection` value (the default is, fill text)
The root cause was that the preexisting calculation simply took the midpoint of the start and end angle, and if it was shorter to go along the other arc of the circle, then that's what it did. So it always resulted in an exact overlap with two slices where one was over 180 degrees and linked labels were forced.

fix #699