fix: set uninitialized state when removeSpec action is called
This commit reset the chart state to not initialized when removing a spec. After #723 PR that
reduced the number of steps on the state machine when parsing, the removeSpec action wasn't
accounted on that refactoring causing side effects when removing/switching a spec on the chart
configuration due to the state being in an wrong status.

fix #738