feat: streamgraph and fit functions on stacked charts (#751)
This commit adds the wiggle and silhouette stacks mode as available in vislib that allow the creation of streamgraph. It is also to fix fit functions for stacked charts. The commit also add the following fixes:
- fit functions are now available also on stacked charts
- the legend extra value is now displayed, on-demand, only on charts with a continuous x-axis
- Adjusts slightly the way fitted data is represented in the case of null values and no fit option specified. It renders an empty area in this case
- for zero-based charts (bar and area charts) the fit option doesn't have any effects
- djusts the clipped ranges for ordinal charts, where the lines where wrongly clipped due to a missing translate call when rendering

BREAKING CHANGE: the first parameter of `PointStyleAccessor` and `BarStyleAccessor` callbacks is changed from `RawDataSeriesDatum` to `DataSeriesDatum`. `stackAsPercentage` prop is replaced by `stackMode` that accept one `StackMode`.

fix #766
fix #715
close #450