feat(bar_chart): add shadow prop for value labels (#785)
Value labels fill can be tweaked with advanced parameters to increase the contrast with the background, using a border parameter. A new textInverted parameter gives also the ability to better respect contrast requirements, enabling an auto tuning of the fill color based on the bar color.

BREAKING CHANGE: The `DisplayValueStyle` `fill` property can now express a border color and width, or let the library pick the best match based on contrast using the textInvertible parameter.