refactor: simplify color manipulation code (#1372)
This PR simplifies and refactor the current color manipulation code. 

BREAKING CHANGES: the `textContrast` `textInvertible` and `textOpacity` in the `Font` type has been removed. This affects the `HeatmapConfig` type in the `cell.label`, `xAxisLabel`, and `yAxisLabel` properties. This also affects the `PartitionConfig` properties: `fillLabel` and `linkLabel`. The `textInvertible` and `textContrast` properties are also removed from the `DisplayValueStyle` used to display bar values labels. The mentioned changes are replaced with the following behavior: always invert the text color depending on the color contrast with its background using pure `black` or `white` colors.