fix(heatmap): snap time bucket to calendar/fixed intervals (#1462)
The commit introduces the ability to specify both calendars and fixed intervals as described in Elasticserach date_histogram aggregation. The fix is required to cover the Lens usage when calendar intervals are used in conjunction with the date_histogram aggs.

BREAKING CHANGE: The `xScaleType` is replaced by the prop `xScale`, which better describes a rasterized time scale with an Elasticsearch compliant interval.