feat(heatmap): allow rotation of x axis labels (#1514)
The `rotation` configuration is available on the `theme.heatmap.xAxisLabel` theme style to rotate the labels of the X-axis from 0 to 90 degrees. The rotation will shrink the chart vertically( and occasionally also horizontally) depending on the length of the labels.
Few changes to the x-axis labels are also landed in this PR:
- categorical labels overflowing the chart boundaries will play a role in shrinking the chart size.
- the X-axis now shows small ticks, center-aligned for categorical scale and left-aligned for time
- the time axis ticks are aligned to the right of the ticks
- time axis labels instead, will not shrink the chart anymore. The number of labels will be reduced gradually to avoid overlaps and overflows. The right edge of the heatmap will always touch the right edge of the rendering area.

BREAKING CHANGE: `width`, `align`, and `baseline` style properties are removed from the `xAxisLabels` and `yAxisLabels` style of the Heatmap theme.

Co-authored-by: Marco Vettorello <vettorello.marco@gmail.com>