feat(annotations): animated fade trigger (#1693)
adds `trigger`/`options` array to allow opting-in to specific animations for a given `Spec` with full control of the animated options.

BREAKING CHANGES: moves the animation configuration from the `AnnotationSpecStyles` to the `AnnotationSpec` and removes existing prop on the styles. Replaces `AnnotationAnimation` type with `AnnotationAnimationConfig` and renames `AnimationConfig` type to `LegacyAnimationConfig` to be removed in future release.