feat(tooltip): composable tooltips - phase 1 (#1725)
Redesigned tooltip styles and adds composable tooltip components for users to develop custom tooltip configurations. Adds the option for `header` and `tooltip` on the default tooltip that provides the tooltip items for contextual headers and footers.

BREAKING CHANGE: `Settings.tooltip` is deprecated in favor of the new `Tooltip` spec. Type changes related to `TooltipProps`, `TooltipSettings` and others see #1725. Changes to tooltip styles from dark to light theme. The Annotation tooltip class `.echAnnotation` used to target the icon, but now refers to the tooltip itself similar to `.echTooltip` and now `.echAnnotation__marker` refers to the icon itself.

```diff
 const tooltipProps = {...};
 <Chart>
-  <Settings tooltip={tooltipProps} />
+  <Tooltip {...tooltipProps} />
 </Chart>
```