feat(metric): add a textual value (#1817)
Adds the ability to display text instead of a number in a metric chart. Enforces better types and requires some, previously optional, properties.

BREAKING CHANGE: the metric types have been changed: the `value: number` and `valueFormatter` are only available for a numerical, progress bar, or trend metric. For the textual metric, the `value` has a `string` type.
The `progressBarDirection` and `trendShape` properties are now required.