feat(legend): add custom legend component (#1889)
The settings now have a new prop `customLegend` which lets users provide there own legend React component that is rendered with the chart state into the legend container.
Analogously to the `customTooltip` in the `Tooltip`, the `customLegend` takes a React component which will receive the state of the standard legend as props. This allows users to provide their own legend implementation. The legend is still rendered inside of the `div.echLegendListContainer`.

Co-authored-by: Marco Vettorello <vettorello.marco@gmail.com>