feat(tooltip): expose tooltip container (#1915)
This commit changes the way tooltip components need to be used. It introduces the `TooltipContainer` container element that needs to be used as the parent element when a custom tooltip is defined using our Tooltip components.

BREAKING CHANGE: You must wrap your `customTooltip` components with a `TooltipContainer` if you are using our `Tooltip` components.