fix(partition): render legend items with zero values  (#1956)
This commit will render the categories with a value of zero in the legend, that were previously omitted.

BREAKING CHANGE: The `Layer.fillColor` function now accepts the following signature: `(key: Key, sortIndex: number, node: ArrayNode, tree: HierarchyOfArrays) => string;` This exposes a similar set of information but allows us more control over the internals to define/generate the legend. The legend for pie/donut/mosaic/treemap charts is sorted in a hierarchical way even if the legend is flat.
The default highlight strategy used to highlight partition elements when hovering over the legend item has been changed to `LegendStrategy.Path`.