chore: update tsconfig structure for src
Newer versions of typescript/vscode now support multiple tsconfig.json
files in a single project. Previously the base tsconfig.json was the only
file used and all others were ignored. Now if a tsconfig.json file is at
the root of a directory that will be applied to all child directories.
This change was causing failures in cli but not is vscode for test files.