fix(tooltip): custom tooltip header context (#1989)
- change header from TooltipValue to PointerValue type
- update tooltip header formatter type
- return full set of values for customTooltips

BREAKING CHANGE: The `header` property of `TooltipInfo` type was simplified to `PointerValue` as to include only relevant properties. This change is propagated to all other types using `header` as a `TooltipValue`. The `TooltipInfo.values` used to conditionally pass only highlighted `TooltipValue`s when using a `customTooltip` and now _always_ passes all `values`.