feat(flame): expose search field text and search text change listener (#2068)
This commit exposes the search text as a prop in the Flame component.
It also exposes a listener onSearchTextChange that is called whenever a change in the search text input provided by the chart happens. It always calls the function with a string, even if the clean button is clicked. There is no difference between passing an empty string or clearing the input, both of these actions result in cleaning the search and the API is anyway cleaner and doesn't need a union type.