feat(xy): render sorting (#2524)
This enables sorting of defined series for ordering in rendering, legend and tooltip for xy charts.

BREAKING CHANGE: The way mixed stacked/nonstacked series are colored now is different from the previous behaviour. Now we color them not by their insert index but by the way we display them in the rendering: from the left to right, bottom top, stacked, nonstacked. This align correctly also the legend colors by default. This does **not** affect colors assigned via a `SeriesColorAccessor`.