fix(flame_chart): fix font size to occupy available space (#2671)
This fixes the font size of text in the flame_chart boxes to occupy the the available space.

This often has been a major hassle when demoing the flamegraph of Universal Profiling on smaller screens, e.g. on a laptop during a conference.
The same issue happens when demoing via zoom on a large screen. The font size just doesn't scale with the browser scaling.